Added endpoint to force re-sync of locally edited plugin
This is a tweak to make a developer's life easier. Say you're working on
a plugin and want to change the schema for it. Before, you would have to
upload a new zip version of the plugin so the Authoring tool would see
the change and recreate the schema. Now, you can force the re-sync but
without uploading a new zip file.

Simply browse to /api/components/sync/:plugin-name and the authoring tool
will reload the config and schema files from the default location (which
is currently /temp/:tenant-id/adapt_framework/src/plugins/:plugin-name